requirements.txt
aiogram
python-dotenv
captcha
